Added mobile topic drawer
Topic drawer is now swipeable on mobile devices.
